eliminate static cast

+++ source/newtonian/two_dimensional/hdf5_diagnostics.cpp
@@ -447,7 +447,7 @@ void WriteDelaunay(Delaunay const& tri, string const& filename)
 	vector<Vector2D> const& cor = tri.getCor();
 	vector<double> x_cor, y_cor;
 	vector<int> facets;
-	int nfacets = tri.get_num_facet();
+	size_t nfacets = tri.get_num_facet();
 
 	H5File file(H5std_string(filename), H5F_ACC_TRUNC);
 
@@ -457,7 +457,7 @@ void WriteDelaunay(Delaunay const& tri, string const& filename)
 		y_cor.push_back(cor[i].y);
 	}
 
-	for (int i = 0; i < nfacets; ++i)
+	for (size_t i = 0; i < nfacets; ++i)
 	{
 	  facets.push_back(static_cast<int>(tri.get_facet(i).vertices.first));
 	  facets.push_back(static_cast<int>(tri.get_facet(i).vertices.second));
